feat(dashboard): Reshape draggable filters behaviour
